Add support for creating contour plots
This commit adds a `withContour` method to `Plot` to add a contour plot
to the plot. The name and colourscheme to be used may be specified as
options. Usage is as follows:

    val p = Plot().withContour(x, y, z, ContourOptions().colorscale("Jet"))

where x, y, and z are Iterables such that z_i is the value to be plotted
at coordinate (x_i, y_i).